2_image/requirements.txt
torch
transformers
torchvision
Pillow
timm 
einops
accelerate>=0.26.0
fastapi
uvicorn